cast `local_scale_tensor` to fp32 for precompute of fp8 dynamic scaling (#713)
* cast local_scale_tensor to fp32


* test, but numerics not looking nice


* `Optional[torch.dtype]`, not `dtype | None`


* Update test/float8/test_fsdp2/test_fsdp2.py

---------

+++ torchao/float8/fsdp_utils.py
@@ -67,7 +67,7 @@ def precompute_float8_dynamic_scale_for_fsdp(module: nn.Module) -> None:
     scale_tensor = torch.finfo(torch.float8_e4m3fn).max / amax_tensor  # Replicate
     if amax_tensor.dtype is torch.float16:
         scale_tensor = torch.clamp(scale_tensor, max=torch.finfo(torch.float16).max)
-    local_scale_tensor = scale_tensor.to_local()
+    local_scale_tensor = scale_tensor.to_local().to(dtype=torch.float32)
     for i, float8_linear in enumerate(float8_linears):
         float8_linear.weight._local_tensor._precomputed_scale = local_scale_tensor[i]
